// modules/recruiting/types/job-submission.ts
export interface JobSubmission {
  id: number;
  title: string;
  description: string;
  city: string;
  state: string;
  postal_code: string | null;
  country_code: string;
  remote_type: string | null;
  company_id: number;
  department_id: number | null;
  recruiter_id: number;
  owner_id: number | null;
  contact_id: number | null;
  portal_id: string | null;
  workflow_id: number | null;
  status_id: number | null;
  status: string;
  category: string | null;
  type: string | null;
  is_hot: boolean;
  start_date: string | null;
  salary: string | null;
  max_rate: string | null;
  duration: string | null;
  openings: number;
  external_id: string | null;
  notes: string | null;
  custom_fields: Array<{ id: string; value: string }>;
  tags: string[];
  user_groups: string[];
  created_by: number;
  reviewed_by: number | null;
  created_at: string;
  updated_at: string;
  published_at: string | null;
  cats_job_id: number | null;
  applications: Array<{ id: number }>;
  job_close_schedule_time: string | null;
  time_zone: string | null;
}